Update evaluate_depth script.
